[IMP] change module name [IMP] readme [FIX] test precommit
