--- mrp_bom_purchase/__manifest__.py
@@ -1,36 +1,0 @@
-# Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
-# @author: Quentin Dupont (quentin.dupont@grap.coop)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-{
-    "name": "MRP BoM Purchase",
-    "summary": "Handle purchase from your Bill of Materials",
-    "version": "16.0.1.0.0",
-    "category": "GRAP - Custom",
-    "author": "GRAP",
-    "website": "https://github.com/grap/grap-odoo-custom",
-    "license": "AGPL-3",
-    "depends": [
-        "mrp",
-        "mrp_bom_product_price_margin",
-    ],
-    "data": [
-        "security/ir.model.access.csv",
-        "data/report_paperformat.xml",
-        "report/report_bom_purchase_list.xml",
-        "report/ir_actions_report.xml",
-        "wizard/view_bom_print_purchase_list_wizard.xml",
-        "views/mrp_bom.xml",
-        "views/action.xml",
-        "views/menu.xml",
-    ],
-    "assets": {
-        "web.report_assets_common": [
-            "mrp_bom_purchase/static/src/scss/mrp_bom_purchase.scss",
-        ],
-        "web.assets_backend": [
-            "mrp_bom_purchase/static/src/scss/mrp_bom_purchase.scss",
-        ],
-    },
-    "installable": True,
-}
