--- mrp_bom_purchase/readme/ROADMAP.rst
@@ -1,3 +1,0 @@
-- Button in wizard to create purchase and manufacturing orders ?
-- How to handle nested of nested boms ? At the moment, take one level into account
-- Field scope bom.description_packaging in this module ?
