--- mrp_bom_purchase/report/ir_actions_report.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
-@author: Quentin DUPONT (quentin.dupont@grap.coop)
-License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-<odoo>
-
-    <record id="bom_purchase_list" model="ir.actions.report">
-        <field name="model">report.bom.purchase.list</field>
-        <field name="name">BoM Purchase List</field>
-        <field name="report_type">qweb-pdf</field>
-        <field name="report_name">mrp_bom_purchase.report_bom_purchase_list</field>
-        <field name="attachment_use">False</field>
-        <field name="paperformat_id" ref="paper_format_purchase_list"/>
-    </record>
-
-</odoo>
