--- mrp_bom_purchase/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_mrp_bom_purchase
