--- mrp_bom_purchase/views/action.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
-@author: Quentin DUPONT (quentin.dupont@grap.coop)
-License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-<odoo>
-
-    <record id="action_mrp_bom_purchase_list_print_wizard" model="ir.actions.act_window">
-        <field name="name">Production assistant</field>
-        <field name="res_model">bom.print.purchase.list.wizard</field>
-        <field name="view_mode">form</field>
-        <field name="target">new</field>
-        <field name="binding_model_id" ref="mrp.model_mrp_bom"/>
-    </record>
-
-</odoo>
