--- mrp_bom_purchase/views/assets.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
-@author: Quentin DUPONT (quentin.dupont@grap.coop)
-License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-<odoo>
-
-    <!-- We add css in assets common, file shared for custom reports -->
-    <template id="mrp_bom_purchase_report_assets" inherit_id="web.report_assets_common">
-        <xpath expr="." position="inside">
-            <link rel="stylesheet" type="text/scss" href="/mrp_bom_purchase/static/src/scss/mrp_bom_purchase.scss"/>
-        </xpath>
-    </template>
-
-    <template id="mrp_bom_purchase_wizard_assets" inherit_id="web.assets_backend">
-        <xpath expr="." position="inside">
-            <link rel="stylesheet" type="text/scss" href="/mrp_bom_purchase/static/src/scss/mrp_bom_purchase_wizard.scss"/>
-        </xpath>
-    </template>
-
-</odoo>
