--- mrp_bom_purchase/views/menu.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
-@author: Quentin DUPONT (quentin.dupont@grap.coop)
-License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
--->
-<odoo>
-
-    <menuitem id="menu_mrp_bom_purchase_list_print"
-        parent="mrp.menu_mrp_root" sequence="50"
-        action="action_mrp_bom_purchase_list_print_wizard"/>
-
-</odoo>
