+++ mrp_bom_wizard_production/__manifest__.py
@@ -1,0 +1,36 @@
+# Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
+# @author: Quentin Dupont (quentin.dupont@grap.coop)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+{
+    "name": "MRP BoM Wizard production",
+    "summary": "Wizard linked to Bill of Materials to help your production.",
+    "version": "16.0.1.0.0",
+    "category": "GRAP - Custom",
+    "author": "GRAP",
+    "website": "https://github.com/grap/grap-odoo-custom",
+    "license": "AGPL-3",
+    "depends": [
+        "mrp",
+        "mrp_bom_product_price_margin",
+    ],
+    "data": [
+        "security/ir.model.access.csv",
+        "data/report_paperformat.xml",
+        "report/report_bom_purchase_list.xml",
+        "report/ir_actions_report.xml",
+        "wizard/view_bom_print_purchase_list_wizard.xml",
+        "views/mrp_bom.xml",
+        "views/action.xml",
+        "views/menu.xml",
+    ],
+    "assets": {
+        "web.report_assets_common": [
+            "mrp_bom_wizard_production/static/src/scss/mrp_bom_wizard_production.scss",
+        ],
+        "web.assets_backend": [
+            "mrp_bom_wizard_production/static/src/scss/mrp_bom_wizard_production_wizard.scss",
+        ],
+    },
+    "installable": True,
+}
