+++ mrp_bom_wizard_production/readme/CONTRIBUTORS.rst
