+++ mrp_bom_wizard_production/readme/ROADMAP.rst
@@ -1,0 +1,4 @@
+- Button in wizard to create purchase and manufacturing orders ?
+- How to handle nested of nested boms ? At the moment, take one level into account
+- Field scope bom.description_packaging in this module ?
+- Fix test04
