+++ mrp_bom_wizard_production/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_bom_print_purchase_list_wizard,access_bom_print_purchase_list_wizard,mrp_bom_wizard_production.model_bom_print_purchase_list_wizard,base.group_user,1,1,1,1
+access_bom_print_purchase_list_wizard_line,access_bom_print_purchase_list_wizard_line,mrp_bom_wizard_production.model_bom_print_purchase_list_wizard_line,base.group_user,1,1,1,1
