+++ mrp_bom_wizard_production/static/src/scss/mrp_bom_wizard_production_wizard.scss
