+++ mrp_bom_wizard_production/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_mrp_bom_wizard_production
