+++ mrp_bom_wizard_production/views/mrp_bom.xml
