+++ mrp_bom_wizard_production/wizard/view_bom_print_purchase_list_wizard.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2022 - Today: GRAP (http://www.grap.coop)
+@author: Quentin DUPONT (quentin.dupont@grap.coop)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<odoo>
+
+    <record id="view_bom_print_purchase_list_wizard_form" model="ir.ui.view">
+        <field name="model">bom.print.purchase.list.wizard</field>
+        <field name="arch" type="xml">
+            <form string="Print Purchase list">
+                <div name="wizard_options" class="wizard_options">
+                  <h5>🔧 Choose options</h5><br/>
+                  <group>
+                    <field name="option_group_by_product_category" widget="boolean_toggle"/>
+                    <field name="option_display_cost" widget="boolean_toggle"/>
+                    <field name="option_print_bom" widget="boolean_toggle"/>
+                  </group>
+                </div>
+                <br/>
+                <div class="Notes">
+                  <h5>📝 Informations on PDF</h5>
+                  <group col="4">
+                    <field name="title_for_pdf" colspan="2"/>
+                    <field name="option_production_date" colspan="2"/>
+                    <field name="notes_for_pdf" placeholder="Order of March 18, 1871 for customer Louise Michel" colspan="2"/>
+                  </group>
+                </div>
+                <br/>
+                <div class="wizard_instrutions">
+                  <h5>🖨️ Choose the quantity of each BoMs</h5><br/>
+                  <field name="no_origin" invisible="1"/>
+                  <group col="12">
+                    <field name="line_ids" nolabel="1" colspan="12">
+                        <tree editable="bottom">
+                            <field name="bom_id" required="1"/>
+                            <field name="currency_id" invisible="1"/>
+                            <field name="bom_origin" attrs="{'column_invisible': [('parent.no_origin', '=', True)]}"/>
+                            <field name="bom_description"/>
+                            <field name="bom_product_qty"/>
+                            <field name="quantity" required="1" widget="numeric_step"/>
+                            <field name="bom_uom_id"/>
+                            <field name="wizard_line_subtotal" sum="Total Cost" widget='monetary' options="{'currency_field': 'currency_id'}" attrs="{'column_invisible': [('parent.option_display_cost', '=', False)]}"/>
+                         </tree>
+                    </field>
+                  </group>
+                </div>
+                <footer>
+                    <button name="print_report" string="Print Production Report" type="object" icon="fa-print" class="oe_highlight" attrs="{'invisible': [('option_print_bom', '!=', False)]}"/>
+                    <button name="print_report" string="Print Production Report and BoMs" type="object" icon="fa-print" class="oe_highlight" attrs="{'invisible': [('option_print_bom', '!=', True)]}"/>
+                    <button icon="fa-ban" special="cancel" string="Cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+</odoo>
