--- setup/mrp_bom_purchase/odoo/addons/mrp_bom_purchase
@@ -1,1 +1,0 @@
-../../../../mrp_bom_purchase
