+++ setup/mrp_bom_wizard_production/odoo/addons/mrp_bom_wizard_production
@@ -1,0 +1,1 @@
+../../../../mrp_bom_wizard_production
