+++ test-requirements.txt
@@ -1,0 +1,1 @@
+git+https://github.com/OCA/manufacture@refs/pull/1227/head#subdirectory=setup/mrp_bom_product_price_margin
